requirements.txt
easydict==1.13
git+https://github.com/oval-group/smooth-topk.git
einops==0.8.1
imageio==2.35.1
Jinja2==3.1.6
matplotlib==3.7.5
numpy==1.24.3
nystrom_attention==0.0.14
openslide_python==1.4.1
openslide_bin==4.0.0.8
openpyxl==3.1.5
pandas==2.0.3
pathlib2==2.3.7.post1
Pillow==11.3.0
pydicom==3.0.1
pytorch_lightning==2.3.0
PyYAML==6.0.2
scikit_learn==1.3.2
seaborn==0.13.2
scikit-image==0.21.0
tensorboardX==2.6.2.2
torch==2.3.0
torchvision==0.18.0
tqdm==4.67.1
umap_learn==0.5.7
